Refactor landing page into section components
